优化用户接口，调整账户金额表的存储

+++ ems-foundation/ems-foundation-user/src/main/java/info/zhihui/ems/foundation/user/service/impl/UserServiceImpl.java
@@ -7,17 +7,14 @@ import info.zhihui.ems.common.exception.BusinessRuntimeException;
 import info.zhihui.ems.common.exception.NotFoundException;
 import info.zhihui.ems.common.paging.PageParam;
 import info.zhihui.ems.common.paging.PageResult;
-import info.zhihui.ems.common.utils.TransactionUtil;
-import cn.dev33.satoken.exception.NotLoginException;
-import cn.dev33.satoken.stp.StpUtil;
-import info.zhihui.ems.components.redis.utils.RedisUtil;
 import info.zhihui.ems.foundation.user.bo.RoleBo;
 import info.zhihui.ems.foundation.user.bo.RoleSimpleBo;
 import info.zhihui.ems.foundation.user.bo.UserBo;
-import info.zhihui.ems.foundation.user.constants.LoginConstant;
 import info.zhihui.ems.foundation.user.dto.*;
 import info.zhihui.ems.foundation.user.entity.UserEntity;
 import info.zhihui.ems.foundation.user.entity.UserRoleEntity;
+import info.zhihui.ems.foundation.user.event.UserDeletedEvent;
+import info.zhihui.ems.foundation.user.event.UserPasswordResetEvent;
 import info.zhihui.ems.foundation.user.event.UserProfileUpdatedEvent;
 import info.zhihui.ems.foundation.user.enums.RoleEnum;
 import info.zhihui.ems.foundation.user.mapper.UserMapper;
@@ -186,14 +183,18 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional(rollbackFor = Exception.class)
     public void delete(@NotNull Integer id) {
-        userRoleRepository.deleteByUserId(id);
-        repository.deleteById(id);
-
-        try {
-            StpUtil.logout(id);
-        } catch (NotLoginException ignore) {
-            // ignore when user is not logged in
+        UserEntity entity = repository.selectById(id);
+        if (entity == null) {
+            throw new NotFoundException("用户不存在或已删除");
         }
+
+        userRoleRepository.deleteByUserId(id);
+        int deleteRows = repository.deleteById(id);
+        if (deleteRows <= 0) {
+            throw new BusinessRuntimeException("删除用户失败");
+        }
+
+        applicationEventPublisher.publishEvent(new UserDeletedEvent(id));
     }
 
     /**
@@ -241,22 +242,7 @@ public class UserServiceImpl implements UserService {
         if (updateRows <= 0) {
             throw new BusinessRuntimeException("重置密码失败");
         }
-        Integer userId = dto.getId();
-        TransactionUtil.afterCommitSyncExecute(() -> clearLoginStateAfterResetPassword(userId));
-    }
-
-    /**
-     * 重置密码成功后，清理登录失败计数并强制用户下线。
-     *
-     * @param userId 用户ID
-     */
-    private void clearLoginStateAfterResetPassword(Integer userId) {
-        try {
-            RedisUtil.deleteObject(LoginConstant.PWD_ERR + userId);
-            StpUtil.logout(userId);
-        } catch (Exception ignore) {
-            // ignore when user is not logged in
-        }
+        applicationEventPublisher.publishEvent(new UserPasswordResetEvent(dto.getId()));
     }
 
     /**
